Refactor some code out of main
