+++ .gitignore
@@ -24,3 +24,5 @@ dist
 .tox/
 
 venv
+
+/local_fonts/
